% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.R
\name{getRecommendations}
\alias{getRecommendations}
\title{getRecommendations}
\usage{
getRecommendations(
  nodes,
  tmdb,
  nRecommendations = 20,
  nlist_threshold = 100,
  nmovie_threshold = 1e+05
)
}
\description{
getRecommendations
}
