% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{searchField}
\alias{searchField}
\title{searchField}
\usage{
searchField(
  pattern,
  movies = movierecommendations::tmdb_movies(),
  field = "unique_title",
  return_first = FALSE
)
}
\description{
searchField
}
